+++ src/main/java/com/project/professor/allocation/service/ProfessorService.java
@@ -7,30 +7,40 @@ import com.project.professor.allocation.repository.ProfessorRepository;
 
 @Service
 public class ProfessorService {
-	
+
 	private final ProfessorRepository professorRepository;
 
 	public ProfessorService(ProfessorRepository professorRepository) {
 		super();
 		this.professorRepository = professorRepository;
 	}
+
 	public Professor findById(Long id) {
 		Professor professor = professorRepository.findById(id).orElse(null);
 		return professor;
 	}
-				
+
 	public Professor create(Professor professor) {
-		professor = professorRepository.save(professor);
-		return professor;
+		professor.setId(null);
+		return professorRepository.save(professor);
 	}
-	
-	public void deleteById (Long id) {
-		professorRepository.deleteById(id);
+
+	public Professor updateProfessor(Professor professor) {
+		Long id = professor.getId();
+		if (id != null && professorRepository.existsById(id)) {
+			return professorRepository.save(professor);
+		}
+		return null;
 	}
-	
-	public void deleteAll (Long id) {
+
+	public void deleteById(Long id) {
+		if (professorRepository.existsById(id)) {
+			professorRepository.deleteById(id);
+		}
+	}
+
+	public void deleteAll(Long id) {
 		professorRepository.deleteAllInBatch();
 	}
-	
-}
 
+}
